Add expr package split from eval_module for multi-file benchmarks.
