+++ upstream/src/minisweagent/run/expr/errors.py
@@ -1,0 +1,4 @@
+class EvalError(Exception):
+    """Error raised during expression evaluation."""
+
+    pass
